fixed expand view
I decided to break up the expand view into different template overloads via the constaint that op type can be fully utilized without constructing an object or passing an already constructed object to the view. These cases will be handled in seperate template overloads to simplify the implementation. This vastly decreased the complexity of the expand view class.